[ADDED] All functionality to detail view.

+++ Project/app/src/main/java/com/csd/huelight/data/LightBulb.java
@@ -6,12 +6,13 @@ public class LightBulb implements Serializable {
     private final String UID;
     private String name;
     private boolean on;
-    private short hue;
-    private byte saturation;
-    private byte brightness;
+    private int hue;
+    private short saturation;
+    private short brightness;
     private boolean colorLoop;
 
-    public LightBulb(String UID, String name, boolean on, short hue, byte saturation, byte brightness, boolean colorLoop) {
+    public LightBulb(String UID, String name, boolean on, int hue, short saturation,
+                     short brightness, boolean colorLoop) {
         this.UID = UID;
         this.name = name;
         this.on = on;
@@ -41,27 +42,27 @@ public class LightBulb implements Serializable {
         this.on = on;
     }
 
-    public short getHue() {
+    public int getHue() {
         return hue;
     }
 
-    public void setHue(short hue) {
+    public void setHue(int hue) {
         this.hue = hue;
     }
 
-    public byte getSaturation() {
+    public short getSaturation() {
         return saturation;
     }
 
-    public void setSaturation(byte saturation) {
+    public void setSaturation(short saturation) {
         this.saturation = saturation;
     }
 
-    public byte getBrightness() {
+    public short getBrightness() {
         return brightness;
     }
 
-    public void setBrightness(byte brightness) {
+    public void setBrightness(short brightness) {
         this.brightness = brightness;
     }
